tests/wpt/tests/html/canvas/offscreen/pixel-manipulation/2d.imageData.object.nan.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.imageData.object.nan</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.imageData.object.nan</h1>
<p class="desc">ImageData.data converts NaN to 0</p>


<script>
var t = async_test("ImageData.data converts NaN to 0");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var imgdata = ctx.getImageData(0, 0, 10, 10);
  imgdata.data[0] = 100;
  imgdata.data[0] = NaN;
  _assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
  imgdata.data[0] = 100;
  imgdata.data[0] = "cheese";
  _assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
  t.done();

});
</script>
